% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basinArm}
\alias{basinArm}
\title{A Basin-Arm table for Sea-bird CTD data}
\format{
A tbl data frame:
\describe{
\item{Lake}{The name of the lake}
\item{BasinArm}{What arms or basins are within each lake}
}
}
\usage{
basinArm
}
\description{
A Basin-Arm table for Sea-bird CTD data
}
\keyword{datasets}
