% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_if_valid_path}
\alias{connect_if_valid_path}
\title{Connect to database if file exists and is and SQLite database
used only if the user provied as path as the conn arg for other exported functions}
\usage{
connect_if_valid_path(path)
}
\arguments{
\item{path}{The path to an SQLite Database}
}
\description{
Connect to database if file exists and is and SQLite database
used only if the user provied as path as the conn arg for other exported functions
}
