% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_metals_init}
\alias{ems_metals_init}
\title{An initialized table for EMS metals data}
\format{
A tbl data frame:
}
\usage{
ems_metals_init
}
\description{
An initialized table for EMS metals data
}
\keyword{datasets}
