% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_param_lookup}
\alias{ems_param_lookup}
\title{A lookup table for for selecting the desired columns from EMS}
\format{A tbl data frame:
\describe{
  \item{PARAMETER_CODE}{PARAMETER_CODE}
  \item{PARAMETER}{PARAMETER}
  \item{ANALYTICAL_METHOD}{ANALYTICAL_METHOD}
  \item{chem.make.names}{chem.make.names}
  \item{chem.names}{chem.names}
  \item{Comment}{Comment}
}}
\usage{
ems_param_lookup
}
\description{
A lookup table for for selecting the desired columns from EMS
}
\keyword{datasets}
