% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_standard_init}
\alias{ems_standard_init}
\title{An initialized table for EMS standard data}
\format{
A tbl data frame:
}
\usage{
ems_standard_init
}
\description{
An initialized table for EMS standard data
}
\keyword{datasets}
