% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emsSites}
\alias{emsSites}
\title{A site table for EMS data}
\format{A tbl data frame:
\describe{
  \item{SiteID}{Site ID}
  \item{EmsSite}{EMS database site name}
  \item{SiteName}{Site name}
}}
\usage{
emsSites
}
\description{
A site table for EMS data
}
\keyword{datasets}
