% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tables.R
\name{initialize_ctd}
\alias{initialize_ctd}
\title{Initialize ctd table for database}
\usage{
initialize_ctd()
}
\value{
initialized table with units
}
\description{
Initialize ctd table for database
}
