% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mysid_params}
\alias{mysid_params}
\title{A character vector of all parameters for the mysid data in the database}
\format{
A character vector
}
\usage{
mysid_params
}
\description{
A character vector of all parameters for the mysid data in the database
}
\keyword{datasets}
