% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrp-package.R
\docType{package}
\name{nrp-package}
\alias{nrp}
\alias{nrp-package}
\title{nrp: Kootenay Lake Nutrient Restoration Program Data Loading}
\description{
Imports CTD, EMS, Zooplankton, and Mysid data for the Kootenay Lake Nutrient Restoration Program (NRP) and provides functionality for uploading and downloading with an SQLite database.
}
\author{
\strong{Maintainer}: Evan Amies-Galonski \email{evanamiesgalonski@gmail.com} (\href{https://orcid.org/0000-0003-1096-2089}{ORCID})

Authors:
\itemize{
  \item Evan Amies-Galonski \email{evanamiesgalonski@gmail.com} (\href{https://orcid.org/0000-0003-1096-2089}{ORCID})
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder, funder]
}

}
\keyword{internal}
