% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_add_ctd_sites}
\alias{nrp_add_ctd_sites}
\title{Add new ctd sites to database site table}
\usage{
nrp_add_ctd_sites(data, db_path)
}
\arguments{
\item{data}{a tibble or data frame of new site data
Must have columns "SiteID", "SiteNumber", "SiteName", "BasinArm", "Depth", as well as columns
"easting" and "northing" with coordinates in projection UTM zone 11N.}

\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\description{
Add new ctd sites to database site table
}
