% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_download_ctd_basin_arm}
\alias{nrp_download_ctd_basin_arm}
\title{Download CTD BasinArm table}
\usage{
nrp_download_ctd_basin_arm(db_path = getOption("nrp.db_path", NULL))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
CTD BasinArm table
}
\description{
Download CTD BasinArm table
}
