% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_download_ctd_sites}
\alias{nrp_download_ctd_sites}
\title{Download CTD site table}
\usage{
nrp_download_ctd_sites(db_path = getOption("nrp.db_path", NULL))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
CTD site table
}
\description{
Download CTD site table
}
