% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_download_ctd_visit}
\alias{nrp_download_ctd_visit}
\title{Download CTD visit table}
\usage{
nrp_download_ctd_visit(db_path = getOption("nrp.db_path", NULL))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
CTD visit table
}
\description{
Download CTD visit table
}
