% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems.R
\name{nrp_download_ems_sites}
\alias{nrp_download_ems_sites}
\title{Download EMS site table}
\usage{
nrp_download_ems_sites(db_path = getOption("nrp.db_path", NULL))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
EMS site table
}
\description{
Download EMS site table
}
