% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_download_lakes}
\alias{nrp_download_lakes}
\title{Download Lake table}
\usage{
nrp_download_lakes(db_path = getOption("nrp.db_path", NULL))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
CTD Lakes table
}
\description{
Download Lake table
}
