% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysid.R
\name{nrp_download_mysid_sample}
\alias{nrp_download_mysid_sample}
\title{Download Mysid sample table}
\usage{
nrp_download_mysid_sample(db_path = getOption("nrp.db_path", file.choose()))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
Mysid sample table
}
\description{
Download Mysid sample table
}
