% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto.R
\name{nrp_download_phyto}
\alias{nrp_download_phyto}
\title{Download Phytoplankton data table from database}
\usage{
nrp_download_phyto(
  start_date = NULL,
  end_date = NULL,
  sites = NULL,
  species = "all",
  db_path = getOption("nrp.db_path", file.choose())
)
}
\arguments{
\item{start_date}{The start date}

\item{end_date}{The end date}

\item{sites}{A character vector of the Site IDs}

\item{species}{A character vector of the species to include. Defaults to 'all'
Permissible values can be found in the PhytoplanktonSpecies table.}

\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
Phytoplankton data table
}
\description{
Download Phytoplankton data table from database
}
