% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto.R
\name{nrp_download_phyto_species}
\alias{nrp_download_phyto_species}
\title{Download phytoplankton species table}
\usage{
nrp_download_phyto_species(db_path = getOption("nrp.db_path", file.choose()))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
The phytoplankton species table
}
\description{
Download phytoplankton species table
}
