% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo.R
\name{nrp_download_zoo_sample}
\alias{nrp_download_zoo_sample}
\title{Download Zooplankton sample table}
\usage{
nrp_download_zoo_sample(db_path = getOption("nrp.db_path", file.choose()))
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
Zooplankton sample table
}
\description{
Download Zooplankton sample table
}
