% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo.R
\name{nrp_download_zooplankton}
\alias{nrp_download_zooplankton}
\title{Download Zooplankton data table from database}
\usage{
nrp_download_zooplankton(
  start_date = NULL,
  end_date = NULL,
  sites = NULL,
  parameters = "all",
  counts = FALSE,
  db_path = getOption("nrp.db_path", file.choose())
)
}
\arguments{
\item{start_date}{The start date}

\item{end_date}{The end date}

\item{sites}{A character vector of the Site IDs}

\item{parameters}{A character vector of the parameters to include.
Permissable values can be found in the nrp::zoo_params}

\item{counts}{A flag indicating whether to return the raw counts instead.}

\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
CTD data table
}
\description{
Download Zooplankton data table from database
}
