% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems.R
\name{nrp_extract_ems}
\alias{nrp_extract_ems}
\title{Extract and clean EMS data}
\usage{
nrp_extract_ems(
  data,
  db_path = getOption("nrp.db_path", file.choose()),
  analysis_type = "standard"
)
}
\arguments{
\item{data}{The EMS data to be extracted. Data must first be downloaded from server by using rems package}

\item{db_path}{The SQLite connection object or path to the nrp SQLite database}

\item{analysis_type}{EMS data of interest. Must be either "standard" or "metals"}
}
\value{
A a data frame
}
\description{
Extract and clean EMS data
}
