% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_read_ctd}
\alias{nrp_read_ctd}
\title{Read CTD File}
\usage{
nrp_read_ctd(path)
}
\arguments{
\item{path}{A string of the path to the file}
}
\value{
A tibble
}
\description{
Read CTD File
}
\examples{
path <- system.file("extdata", "ctd/2018/KL1_27Aug2018008downcast.cnv",
             package = "nrp")
nrp_read_ctd(path)
}
