% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto.R
\name{nrp_read_phyto_file}
\alias{nrp_read_phyto_file}
\title{Read phytoplankton raw data file}
\usage{
nrp_read_phyto_file(path, db_path = getOption("nrp.db_path", file.choose()))
}
\arguments{
\item{path}{A string of the path to the file.}

\item{db_path}{The SQLite connection object or path to the nrp SQLite database.}
}
\value{
A tibble
}
\description{
Read phytoplankton raw data file
}
