% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo.R
\name{nrp_read_zooplankton_file}
\alias{nrp_read_zooplankton_file}
\title{Read zooplankton raw data file}
\usage{
nrp_read_zooplankton_file(
  path,
  db_path = getOption("nrp.db_path", file.choose()),
  system = NULL
)
}
\arguments{
\item{path}{A string of the path to the file.}

\item{db_path}{The SQLite connection object or path to the nrp SQLite database.}

\item{system}{The system 'arrow' or 'kootenay'. If null, the system is detected from the file name.}
}
\value{
A tibble
}
\description{
Read zooplankton raw data file
}
