% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{nrp_upload}
\alias{nrp_upload}
\title{Upload data to nrp database}
\usage{
nrp_upload(data, conn)
}
\arguments{
\item{data}{the object name of the data to be uploaded}

\item{conn}{A formal SQlite connection}
}
\description{
Upload data to nrp database
}
