% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyto_input_cols}
\alias{phyto_input_cols}
\title{A named character vector with the correct column names, data types, and order for raw phytoplankton data}
\format{
A character vector:
}
\usage{
phyto_input_cols
}
\description{
A named character vector with the correct column names, data types, and order for raw phytoplankton data
}
\keyword{datasets}
