% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyto_species}
\alias{phyto_species}
\title{A reference table for phytoplankton species}
\format{
A tbl data frame:
\describe{
  \item{Taxa}{The species or sub-species}
  \item{Genus}{The genus}
  \item{ClassName}{The formal class name(s)}
  \item{ClassAlias}{The informal class name(s)}
}
}
\usage{
phyto_species
}
\description{
A reference table for phytoplankton species
}
\keyword{datasets}
