% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_date_lookup}
\alias{site_date_lookup}
\title{A lookup table fr matching historical file names with Site IDs and Dates}
\format{A tbl data frame:
\describe{
  \item{File}{File name}
  \item{Date}{Date data was collected}
  \item{SiteID}{Site ID}
}}
\usage{
site_date_lookup
}
\description{
A lookup table fr matching historical file names with Site IDs and Dates
}
\keyword{datasets}
