% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites}
\alias{sites}
\title{A site table for Sea-bird CTD data}
\format{A tbl data frame:
\describe{
  \item{SiteID}{Site Identifyer}
  \item{SiteNumber}{Site Number}
  \item{SiteName}{Common Name or description the Site}
  \item{BasinArm}{Which arm or Basin it is located in}
  \item{Depth}{The depth of the Seabird}
  \item{geometry}{Coordinates for the lccation of the site}
}}
\usage{
sites
}
\description{
A site table for Sea-bird CTD data
}
\keyword{datasets}
