% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoo_cols}
\alias{zoo_cols}
\title{A character vector with the correct column names and order for raw zooplankton data}
\format{
A character vector:
}
\usage{
zoo_cols
}
\description{
A character vector with the correct column names and order for raw zooplankton data
}
\keyword{datasets}
