% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoo_input_cols}
\alias{zoo_input_cols}
\title{A named character vector with the correct column names, data types,  and order for raw zooplankton data}
\format{
A character vector:
}
\usage{
zoo_input_cols
}
\description{
A named character vector with the correct column names, data types,  and order for raw zooplankton data
}
\keyword{datasets}
