% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoo_params}
\alias{zoo_params}
\title{A character vector of all parameters for the zooplankton data in the database}
\format{
A character vector:
}
\usage{
zoo_params
}
\description{
A character vector of all parameters for the zooplankton data in the database
}
\keyword{datasets}
