% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lakes}
\alias{lakes}
\title{A lakes table for Sea-bird CTD data}
\format{
A tbl data frame:
\describe{
  \item{Lake}{The name of the lake}
  \item{Area}{the area of te lake in meters squared}
  \item{geometry}{Spatial polygon data}
}
}
\usage{
lakes
}
\description{
A lakes table for Sea-bird CTD data
}
\keyword{datasets}
