% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mysid_input_cols}
\alias{mysid_input_cols}
\title{A named character vector with the correct column names, data types, and order for raw mysid data}
\format{
A character vector:
}
\usage{
mysid_input_cols
}
\description{
A named character vector with the correct column names, data types, and order for raw mysid data
}
\keyword{datasets}
