% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrp-package.R
\docType{package}
\name{nrp-package}
\alias{nrp}
\alias{nrp-package}
\title{nrp: Nutrient Restoration Program Data Loading}
\description{
nrp is an R package that is being developed for the 
    Nutrient Restoration Program (NRP) to read the various data files 
    used by the program; upload data to the NRP SQLite database; and 
    produce key tables and plots based on the data in the NRP SQLite database.
}
\author{
\strong{Maintainer}: Evan Amies-Galonski \email{evan@poissonconsulting.ca} (-3188)

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)
}

}
\keyword{internal}
