% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrp-package.R
\docType{package}
\name{nrp-package}
\alias{nrp}
\alias{nrp-package}
\title{nrp: Kootenay Lake Nutrient Restoration Program Data Loading}
\description{
Imports CTD and EMS data for 
    the Kootenay Lake Nutrient Restoration Program (NRP) and uploads it to to a SQLite database.
}
\author{
\strong{Maintainer}: Evan Amies-Galonski \email{evan@poissonconsulting.ca} (-3188)

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)
}

}
\keyword{internal}
