% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_download_ctd}
\alias{nrp_download_ctd}
\title{Download CTD data table from database}
\usage{
nrp_download_ctd(
  start_date = NULL,
  end_date = NULL,
  sites = NULL,
  parameters = "all",
  db_path = getOption("nrp.db_path", file.choose())
)
}
\arguments{
\item{start_date}{The start date}

\item{end_date}{The end date}

\item{sites}{A character vector of the Site IDs}

\item{parameters}{A character vector of the parameters to include.
Permissable values: "Temperature", "Oxygen", "Oxygen2", "Conductivity",
"Conductivity2", "Salinity", "Backscatter", "Fluorescence", "Frequency",
"Flag", "Pressure"}

\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
CTD data table
}
\description{
Download CTD data table from database
}
