% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems.R
\name{nrp_download_ems}
\alias{nrp_download_ems}
\title{Dowload EMS data table from database}
\usage{
nrp_download_ems(
  db_path = getOption("nrp.db_path", file.choose()),
  start_date_time = NULL,
  end_date_time = NULL,
  sites = NULL,
  analysis_type = "standard",
  show_detection_limits = FALSE
)
}
\arguments{
\item{db_path}{The SQLite connection object or path to the SQLite database}

\item{start_date_time}{The start date}

\item{end_date_time}{The end date}

\item{sites}{A character vector of the Site IDs}

\item{analysis_type}{EMS data of interest. Must be either "standard" or "metals"}

\item{show_detection_limits}{Whether to include detection limit columns for each parameter}
}
\value{
EMS data table
}
\description{
Dowload EMS data table from database
}
