% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_load_ctd}
\alias{nrp_load_ctd}
\title{Load CTD data table from database}
\usage{
nrp_load_ctd(start_date = NULL, end_date = NULL, sites = NULL,
  db_path = getOption("nrp.db_path", NULL))
}
\arguments{
\item{start_date}{The start date}

\item{end_date}{The end date}

\item{sites}{A character vector of the Site IDs}

\item{db_path}{The SQLite connection object or path to the SQLite database}
}
\value{
CTD data table
}
\description{
Load CTD data table from database
}
