% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_read_ctd}
\alias{nrp_read_ctd}
\title{Read CTD Files}
\usage{
nrp_read_ctd(path = ".", db_path = getOption("nrp.db_path", NULL),
  recursive = FALSE, regexp = "[.]cnv$", fail = TRUE)
}
\arguments{
\item{path}{A string of the path to the directory.}

\item{db_path}{The SQLite connection object or path to the SQLite database}

\item{recursive}{Should directories be listed recursively?
the filenames.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{fail}{Should the call fail (the default) or warn if a file cannot be
accessed.}
}
\value{
A list of tibbles.
}
\description{
Read CTD Files
}
