% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{nrp_read_ctd}
\alias{nrp_read_ctd}
\title{Read CTD Files}
\usage{
nrp_read_ctd(path = ".", recursive = FALSE, bind = TRUE,
  regexp = "[.]cnv$", fail = TRUE)
}
\arguments{
\item{path}{A string of the path to the directory.}

\item{recursive}{Should directories be listed recursively?
the filenames.}

\item{bind}{A logical argument for whether to bind rows into one tibble or return as a list of tibbles}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{fail}{Should the call fail (the default) or warn if a file cannot be
accessed.}
}
\value{
A list of tibbles.
}
\description{
Read CTD Files
}
\examples{
path <- system.file("extdata", "ctd/2018", package = "nrp")
nrp_read_ctd(path)
}
